Inline the request type modules
this is better for separate compilation.

For now we don't inline the output modules because they may be shared
across different operations.

This also allows us to get rid of the types_1 / types_2 hack for EC2
that caused a stack overflow in the compiler now that the Types module
is smaller.